@@ -1,32 +1,24 @@
 {
-  "name": "rhysbrooker.com",
-  "description": "Gatsby Static site",
+  "name": "gatsby-starter-default",
+  "description": "Gatsby default starter",
   "version": "1.0.0",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/105ron/portfoliosite.git"
-  },
   "dependencies": {
     "babel-plugin-styled-components": "^1.7.1",
-    "gatsby": "^2.0.6",
-    "gatsby-image": "^2.0.7",
-    "gatsby-link": "^2.0.1",
+    "gatsby": "^2.0.0",
+    "gatsby-image": "^2.0.8",
+    "gatsby-plugin-manifest": "^2.0.2",
+    "gatsby-plugin-offline": "^2.0.5",
     "gatsby-plugin-react-helmet": "^3.0.0",
     "gatsby-plugin-sharp": "^2.0.5",
     "gatsby-plugin-styled-components": "^3.0.0",
-    "gatsby-remark-prismjs": "^3.0.0",
-    "gatsby-source-contentful": "^2.0.1",
     "gatsby-source-filesystem": "^2.0.1",
-    "gatsby-transformer-remark": "^2.1.3",
     "gatsby-transformer-sharp": "^2.1.1",
-    "global": "^4.3.2",
-    "hoek": "^5.0.4",
+    "install": "^0.12.1",
     "npm": "^6.4.1",
-    "prismjs": "^1.15.0",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "react-google-recaptcha": "^1.0.2",
+    "prop-types": "^15.6.2",
+    "react": "^16.5.1",
+    "react-dom": "^16.5.1",
     "react-helmet": "^5.2.0",
     "styled-components": "^3.4.9"
   },
@@ -34,22 +26,21 @@
     "gatsby"
   ],
   "license": "MIT",
-  "main": "n/a",
   "scripts": {
     "build": "gatsby build",
     "develop": "gatsby develop",
-    "format": "prettier --trailing-comma es5 --no-semi --single-quote --write \"src/**/*.js\"",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "deploy": "gatsby build --prefix-paths && gh-pages -d public"
+    "format": "prettier --write '**/*.js'",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/gatsbyjs/gatsby-starter-default"
   },
   "devDependencies": {
-    "dotenv": "^6.0.0",
-    "eslint": "^4.19.1",
-    "eslint-config-airbnb": "^17.0.0",
-    "eslint-plugin-import": "^2.13.0",
-    "eslint-plugin-jsx-a11y": "^6.1.0",
-    "eslint-plugin-react": "^7.10.0",
-    "gh-pages": "^2.0.0",
-    "prettier": "^1.14.3"
+    "eslint": "^5.6.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.1",
+    "eslint-plugin-react": "^7.11.1"
   }
 }
