@@ -1,35 +1,7 @@
-const path = require('path')
+/**
+ * Implement Gatsby's Node APIs in this file.
+ *
+ * See: https://www.gatsbyjs.org/docs/node-apis/
+ */
 
-exports.createPages = ({graphql, boundActionCreators}) => {
-  const {createPage} = boundActionCreators
-  return new Promise((resolve, reject) => {
-    const blogPostTemplate = path.resolve('src/templates/blog-post.js')
-    resolve(
-      graphql(`
-        {
-          allContentfulBlog (limit: 100) {
-            edges {
-              node {
-                slug
-              }
-            }
-          }
-        }
-      `).then((result) => {
-        if (result.errors) {
-          reject(result.errors)
-        }
-        result.data.allContentfulBlog.edges.forEach((edge) => {
-          createPage ({
-            path: edge.node.slug,
-            component: blogPostTemplate,
-            context: {
-              slug: edge.node.slug
-            }
-          })
-        })
-        return
-      })
-    )
-  })
-}
+// You can delete this file if you're not using it
