@@ -1,13 +1,8 @@
-import SocialIcons from './socialIcons';
-import icon48 from './favicon48.png';
-import icon72 from './favicon72.png';
-import icon96 from './favicon96.png';
+import {
+  Email, GitHub, LinkedIn, Twitter,
+} from './socialIcons';
 
-const {
-  Email, GitHub, Linkedin, Twitter,
-} = SocialIcons;
-
-const formFields = [
+export const formFields = [
   {
     label: 'First Name:',
     type: 'name',
@@ -34,7 +29,7 @@ const formFields = [
   },
 ];
 
-const iconsAndLinks = [
+export const iconsAndLinks = [
   {
     icon: Email,
     link: `mailto:rhysbrooker01@gmail.com?subject=Email%20from%20RhysBrooker.com`,
@@ -44,7 +39,7 @@ const iconsAndLinks = [
     link: `https://github.com/105ron`,
   },
   {
-    icon: Linkedin,
+    icon: LinkedIn,
     link: `https://www.linkedin.com/in/rhysbrooker/`,
   },
   {
@@ -53,32 +48,7 @@ const iconsAndLinks = [
   },
 ];
 
-const manifest = {
-  name: "MadeByRhys",
-  short_name: "RhysB",
-  start_url: ".",
-  display: "standalone",
-  background_color: "#fff",
-  description: "A portfolio site, blog and projects of Rhys.",
-  icons: [{
-    rel: "icon",
-    href: `${icon48}`,
-    sizes: "48x48",
-    type: "image/png",
-  }, {
-    rel: "icon",
-    href: `${icon72}`,
-    sizes: "72x72",
-    type: "image/png",
-  }, {
-    rel: "icon",
-    href: `${icon96}`,
-    sizes: "96x96",
-    type: "image/png",
-  }],
-};
-
-const pages = [
+export const pages = [
   {
     name: 'Home',
     route: '/',
@@ -101,19 +71,10 @@ const pages = [
   },
 ];
 
-const pageObject = {
+export const pageObject = {
   home: pages[0],
   about: pages[1],
   blogs: pages[2],
   projects: pages[3],
   contact: pages[4],
-};
-
-
-module.exports = {
-  formFields,
-  iconsAndLinks,
-  manifest,
-  pages,
-  pageObject,
 };
